nav: moved addWallet to separate stack as a modal
We want to be able to pop off all the modals and revert to the home screen after the wallet has been created, and the user indicates they are done backing up the mnemonic.

In general, the rule of thumb is to have modal stacks for screens that represent temporary actions and not part of the main wallet's UI, i.e. creating or restoring a wallet, backup page, etc.